+++ main.py
@@ -5,6 +5,7 @@
 Player class for Raspberry Pi3. Can set up audio sink and play/pause/stop the reproducing of WAV files
 
 Changelogs:
+1.8.1 - COME TOGETHER - Changed ready prompt
 1.8.0 - COME TOGETHER - Third stable release: added usb update
 1.7.0 - BLACKBIRD - second stable release: added bluetooth unpairing function/enhanced performances
 1.6.0 - ACROSS THE UNIVERSE - first stable release