+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Come Together - 1.8.1] - 04-12-2023
+
+### Changed
+- Ready prompt
+
 ## [Come Together - 1.8.0] - 04-12-2023
 
 Third release: implemented update over usb